Fixed typo on chpasswd.

+++ configure-server.bash
@@ -9,15 +9,20 @@ set +a
 
 # Firewall
 apt-get install ufw -y
-ufw --force reset
-ufw logging on
-ufw logging full
-ufw default allow outgoing
-ufw default deny incoming
-ufw default deny routed
-ufw allow in proto tcp from any to any port "$ENV_SSH_PORT"
-ufw enable
-ufw status verbose | grep -q 'Status: active'
+if ufw status | grep -q 'Status: active' \
+    && ufw status numbered | grep -q -P '\[ 1] 4242/tcp.*ALLOW' \
+    && ufw status numbered | grep -q -P '\[ 2] 4242/tcp.*ALLOW';
+then
+    ufw --force reset
+    ufw logging on
+    ufw logging full
+    ufw default allow outgoing
+    ufw default deny incoming
+    ufw default deny routed
+    ufw allow in proto tcp from any to any port "$ENV_SSH_PORT"
+    ufw enable
+    ufw status verbose | grep -q 'Status: active'
+fi
 
 # Check AppArmor
 apt-get install --yes apparmor
@@ -32,7 +37,9 @@ apt-get install --yes ssh
 if ! systemctl status --no-pager sshd; then
     systemctl enable ssh
 fi
-cp "/etc/ssh/sshd_config" "/etc/ssh/sshd_config.old"
+if [[ ! -f "/etc/ssh/sshd_config.old" ]]; then
+    cp "/etc/ssh/sshd_config" "/etc/ssh/sshd_config.old"
+fi
 cat > "/etc/ssh/sshd_config" << EOF 
 AcceptEnv LANG LC_*
 X11Forwarding no
@@ -53,9 +60,9 @@ systemctl restart ssh
 # pwquality
 apt-get install --yes libpam-pwquality
 
-if [[ ! -e "/etc/security/pwquality.conf.old" ]]; then
+if [[ ! -f "/etc/security/pwquality.conf.old" ]]; then
     cp "/etc/security/pwquality.conf" "/etc/security/pwquality.conf.old" 
-
+fi
 cat << EOF > "/etc/security/pwquality.conf"
 difok = 7
 minlen = 10
@@ -74,7 +81,7 @@ sed -i "/^PASS_WARN_AGE/s/.*/PASS_WARN_AGE\t7" "/etc/login.defs"
 
 # Add non-root user
 adduser dmontesd42
-chpasswd "dmontesd42:${ENV_USER_PASSWORD}"
+echo "dmontesd42:${ENV_USER_PASSWORD}" | chpasswd
 
 # Remove script from .profile
 sed -i '/\/root\/born2beroot\/configure-server.bash/d' "/root/born2beroot/.profile"
